server files updated and divided into db,models,controllers,routes and index.js
