+++ server/src/routes/notificationRoute.js
@@ -1,0 +1,19 @@
+const express = require("express");
+
+const router = express.Router();
+
+const {
+  getNotifications,
+  readUnreadNotifications,
+} = require("../controllers/notificationController");
+
+// getNotifications route
+router.get("/users/:userId/notifications", getNotifications);
+
+// readUnreadNotifications route
+router.put(
+  "/users/:userId/notifications/:notificationId/read",
+  readUnreadNotifications
+);
+
+module.exports = router;
